Move PromoteBranch into DatabaseBranchesService.
